pyproject.toml
[project]
name = "sonality"
version = "0.1.0"
description = "LLM agent with a self-evolving personality via the Sponge architecture"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "datasets>=4.6.1",
    "neo4j>=5.0.0",
    "pgvector>=0.3.0",
    "psycopg[binary,pool]>=3.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "mypy>=1.0.0",
    "playwright>=1.58.0",
]
docs = [
    "zensical>=0.0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
sonality = "sonality.cli:main"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["sonality", "tests", "benches"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["sonality"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "live: tests/benches that require external model API access",
    "bench: evaluation benchmark suites",
]
